requirements.txt
azure-functions
azure-functions-durable==1.3.2
azure-storage-blob==12.25.1
azure-ai-formrecognizer==3.3.2
requests
